[V26-968] Split daily close helper modules
