+++ packages/athena-webapp/docs/agent/test-index.md
@@ -93,6 +93,9 @@ This index enumerates the current automated test files and ties them back to the
 - [`convex/operations/approvalRequests.test.ts`](../../convex/operations/approvalRequests.test.ts)
 - [`convex/operations/customerProfiles.test.ts`](../../convex/operations/customerProfiles.test.ts)
 - [`convex/operations/dailyClose.test.ts`](../../convex/operations/dailyClose.test.ts)
+- [`convex/operations/dailyClose/adjustmentReports.test.ts`](../../convex/operations/dailyClose/adjustmentReports.test.ts)
+- [`convex/operations/dailyClose/approval.test.ts`](../../convex/operations/dailyClose/approval.test.ts)
+- [`convex/operations/dailyClose/automationPolicy.test.ts`](../../convex/operations/dailyClose/automationPolicy.test.ts)
 - [`convex/operations/dailyManagerReportEmail.test.ts`](../../convex/operations/dailyManagerReportEmail.test.ts)
 - [`convex/operations/dailyOpening.test.ts`](../../convex/operations/dailyOpening.test.ts)
 - [`convex/operations/dailyOperations.test.ts`](../../convex/operations/dailyOperations.test.ts)
